requirements.txt
azure-identity
azure-mgmt-resource
pyyaml